Move add-base-babel-config.ts
