+++ src/main/setup/setup.ts
@@ -3,12 +3,12 @@ import { capitalizeFirstLetter } from "../helpers/capitalize-first-letter"
 import { logInfo } from "../logging"
 import { printFinalMessages } from "./print-final-messages"
 import { Step } from "./step"
-import { addBaseBabelConfigStep } from "./steps/add-base-babel-config"
 import { addBaseTestScriptStep } from "./steps/add-base-test-script"
 import { addContentStep } from "./steps/add-content/add-content"
 import { addGitAttributesStep } from "./steps/add-git-attributes"
 import { addGithubWorkflowStep } from "./steps/add-github-workflow"
 import { addReadmeStep } from "./steps/add-readme/add-readme"
+import { addBaseBabelConfigStep } from "./steps/babel/add-base-babel-config"
 import { copyAssetsStep } from "./steps/copy-assets"
 import { createNextAppStep } from "./steps/create-next-app"
 import { formatProjectStep } from "./steps/format-project"
@@ -18,11 +18,11 @@ import { installFramerMotionStep } from "./steps/install-framer-motion"
 import { installReactHookFormStep } from "./steps/install-react-hook-form"
 import { removeOfficialCNAContentStep } from "./steps/remove-official-cna-content"
 import { setUpChakraUIStep } from "./steps/set-up-chakra-ui"
-import { setUpMaterialUIStep } from "./steps/set-up-material-ui"
 import { setUpCssModulesWithSassStep } from "./steps/set-up-css-modules-with-sass"
 import { setUpEmotionStep } from "./steps/set-up-emotion"
 import { setUpEslintStep } from "./steps/set-up-eslint"
 import { setUpLintStagedStep } from "./steps/set-up-lint-staged"
+import { setUpMaterialUIStep } from "./steps/set-up-material-ui"
 import { setUpPrettierStep } from "./steps/set-up-prettier"
 import { setUpStyledComponentsStep } from "./steps/set-up-styled-components"
 import { updateYarnStep } from "./steps/update-yarn"
